Split st7789.py, create displaycore.py
