+++ src/lib/display/st7789.py
@@ -64,6 +64,7 @@ from time import sleep_ms
 import framebuf
 
 from .palette import Palette
+from .displaycore import DisplayCore
 
 
 # mh_if frozen:
@@ -134,38 +135,37 @@ _BIT0 = const(0x01)
 # fmt: off
 
 # Rotation tables
-#   (madctl, width, height, xstart, ystart, needs_swap)[rotation % 4]
+#   (madctl, width, height, xstart, ystart)[rotation % 4]
 
 _DISPLAY_240x320 = const((
-    (0x00, 240, 320, 0, 0, False),
-    (0x60, 320, 240, 0, 0, False),
-    (0xc0, 240, 320, 0, 0, False),
-    (0xa0, 320, 240, 0, 0, False)))
+    (0x00, 240, 320, 0, 0),
+    (0x60, 320, 240, 0, 0),
+    (0xc0, 240, 320, 0, 0),
+    (0xa0, 320, 240, 0, 0)))
 
 _DISPLAY_240x240 = const((
-    (0x00, 240, 240,  0,  0, False),
-    (0x60, 240, 240,  0,  0, False),
-    (0xc0, 240, 240,  0, 80, False),
-    (0xa0, 240, 240, 80,  0, False)))
+    (0x00, 240, 240,  0,  0),
+    (0x60, 240, 240,  0,  0),
+    (0xc0, 240, 240,  0, 80),
+    (0xa0, 240, 240, 80,  0)))
 
 _DISPLAY_135x240 = const((
-    (0x00, 135, 240, 52, 40, False),
-    (0x60, 240, 135, 40, 53, False),
-    (0xc0, 135, 240, 53, 40, False),
-    (0xa0, 240, 135, 40, 52, False)))
+    (0x00, 135, 240, 52, 40),
+    (0x60, 240, 135, 40, 53),
+    (0xc0, 135, 240, 53, 40),
+    (0xa0, 240, 135, 40, 52)))
 
 _DISPLAY_128x128 = const((
-    (0x00, 128, 128, 2, 1, False),
-    (0x60, 128, 128, 1, 2, False),
-    (0xc0, 128, 128, 2, 1, False),
-    (0xa0, 128, 128, 1, 2, False)))
+    (0x00, 128, 128, 2, 1),
+    (0x60, 128, 128, 1, 2),
+    (0xc0, 128, 128, 2, 1),
+    (0xa0, 128, 128, 1, 2)))
 
 # index values into rotation table
 _WIDTH = const(0)
 _HEIGHT = const(1)
 _XSTART = const(2)
 _YSTART = const(3)
-_NEEDS_SWAP = const(4)
 
 # Supported displays (physical width, physical height, rotation table)
 _SUPPORTED_DISPLAYS = const((
@@ -201,7 +201,7 @@ _ST7789_INIT_CMDS = const((
 
 
 
-class ST7789:
+class ST7789(DisplayCore):
     """ST7789 driver class.
 
     Args:
@@ -261,9 +261,7 @@ class ST7789:
         color_order='BGR',
         custom_init=None,
         custom_rotations=None,
-        reserved_bytearray = None,
-        use_tiny_buf = False,
-        **kwargs,  # noqa: ARG002
+        **kwargs,
     ):
         """Initialize display."""
         self.rotations = custom_rotations or self._find_rotations(width, height)
@@ -278,38 +276,8 @@ class ST7789:
             msg = "dc pin is required."
             raise ValueError(msg)
 
-        #init the fbuf
-        if reserved_bytearray is None:
-            # use_tiny_fbuf tells us to use a smaller framebuffer (4 bits per pixel rather than 16 bits)
-            if use_tiny_buf:
-                # round width up to 8 bits
-                size = (height * width) // 2 if (width % 8 == 0) else (height * (width + 1)) // 2
-                reserved_bytearray = bytearray(size)
+        super().__init__(width, height, rotation, **kwargs)
 
-            else: # full sized buffer
-                reserved_bytearray = bytearray(height*width*2)
-
-        self.fbuf = framebuf.FrameBuffer(
-            reserved_bytearray,
-            # height and width are swapped when rotation is 1 or 3
-            height if (rotation % 2 == 1) else width,
-            width if (rotation % 2 == 1) else height,
-            # use_tiny_fbuf uses GS4 format for less memory usage
-            framebuf.GS4_HMSB if use_tiny_buf else framebuf.RGB565,
-            )
-
-        self.palette = Palette()
-        self.palette.use_tiny_buf = self.use_tiny_buf = use_tiny_buf
-
-        # keep track of min/max y vals for writing to display
-        # this speeds up drawing significantly.
-        # only y value is currently used, because framebuffer is stored in horizontal lines,
-        # making y slices much simpler than x slices.
-        self._show_y_min = width if (rotation % 2 == 1) else height
-        self._show_y_max = 0
-
-        self.width = width
-        self.height = height
         self.xstart = 0
         self.ystart = 0
         self.spi = spi
@@ -325,7 +293,6 @@ class ST7789:
         self.init(init_cmds)
         self.init(init_cmds)
         self.rotation(self._rotation)
-        self.needs_swap = True
         self.fill(0x0)
         self.show()
 
@@ -355,34 +322,7 @@ class ST7789:
             sleep_ms(delay)
 
 
-    def _reset_show_min(self):
-        """Reset show boundaries"""
-        self._show_y_min = self.height
-        self._show_y_max = 0
 
-
-    @micropython.viper
-    def _set_show_min(self, y0:int, y1:int):
-        """Set/store minimum and maximum Y to show next time show() is called."""
-        y_min = int(self._show_y_min)
-        y_max = int(self._show_y_max)
-        
-        
-        if y_min > y0:
-            y_min = y0
-        if y_max < y1:
-            y_max = y1
-
-        self._show_y_min = y_min
-        self._show_y_max = y_max
-
-    
-    @micropython.viper
-    def _format_color(self, color:int) -> int:
-        """Swap color bytes if needed, do nothing otherwise."""
-        if (not self.use_tiny_buf) and self.needs_swap:
-            color = ((color & 0xff) << 8) | (color >> 8)
-        return color
 
 
     def _write(self, command=None, data=None):
@@ -400,7 +340,7 @@ class ST7789:
     
     
     @micropython.viper
-    def _write_tiny_buf(self):
+    def _write_tiny_buf(self, y_min: int, y_max: int):
         """Convert tiny_buf data to RGB565 and write to SPI"""
         if self.cs:
             self.cs.off()
@@ -411,8 +351,8 @@ class ST7789:
         height = int(self.height)
         width = int(self.width)
         
-        start_y = int(self._show_y_min)
-        end_y = int(self._show_y_max)
+        start_y = int(y_min)
+        end_y = int(y_max)
         
         # swap colors if needed
         if self.needs_swap:
@@ -462,10 +402,10 @@ class ST7789:
         return output_buf
 
 
-    def _write_normal_buf(self):
+    def _write_normal_buf(self, y_min, y_max):
         """Write normal framebuf data, respecting show_y_min/max values."""
-        source_start_idx = self._show_y_min * self.width * 2
-        source_end_idx = self._show_y_max * self.width * 2
+        source_start_idx = y_min * self.width * 2
+        source_end_idx = y_max * self.width * 2
         
         if source_start_idx < source_end_idx:
             self._write(None, memoryview(self.fbuf)[source_start_idx:source_end_idx])
@@ -507,15 +447,15 @@ class ST7789:
             mode
         """
         # mh_if TDECK:
-        # TDeck shares SPI with SDCard
-        self.spi.init()
+        # # TDeck shares SPI with SDCard
+        # self.spi.init()
         # mh_end_if
         if value:
             self._write(_ST7789_SLPIN)
         else:
             self._write(_ST7789_SLPOUT)
         # mh_if TDECK:
-        self.spi.deinit()
+        # self.spi.deinit()
         # mh_end_if
 
 
@@ -528,15 +468,15 @@ class ST7789:
             inversion mode
         """
         # mh_if TDECK:
-        # TDeck shares SPI with SDCard
-        self.spi.init()
+        # # TDeck shares SPI with SDCard
+        # self.spi.init()
         # mh_end_if
         if value:
             self._write(_ST7789_INVON)
         else:
             self._write(_ST7789_INVOFF)
         # mh_if TDECK:
-        self.spi.deinit()
+        # self.spi.deinit()
         # mh_end_if
 
 
@@ -561,7 +501,6 @@ class ST7789:
             self.height,
             self.xstart,
             self.ystart,
-            self.needs_swap,
         ) = self.rotations[rotation]
 
         if self.color_order == _BGR:
@@ -594,511 +533,37 @@ class ST7789:
             self._write(_ST7789_RAMWR)
 
 
-    def vline(self, x, y, length, color):
-        """
-        Draw vertical line at the given location and color.
-
-        Args:
-            x (int): x coordinate
-            Y (int): y coordinate
-            length (int): length of line
-            color (int): 565 encoded color
-        """
-        self._set_show_min(y, y + length)
-        color = self._format_color(color)
-        self.fbuf.vline(x, y, length, color)
-
-
-    def hline(self, x, y, length, color):
-        """
-        Draw horizontal line at the given location and color.
-
-        Args:
-            x (int): x coordinate
-            Y (int): y coordinate
-            length (int): length of line
-            color (int): 565 encoded color
-        """
-        self._set_show_min(y, y)
-        color = self._format_color(color)
-        self.fbuf.hline(x, y, length, color)
-
-
-    def pixel(self, x, y, color):
-        """
-        Draw a pixel at the given location and color.
-
-        Args:
-            x (int): x coordinate
-            Y (int): y coordinate
-            color (int): 565 encoded color
-        """
-        self._set_show_min(y, y)
-        color = self._format_color(color)
-        self.fbuf.pixel(x,y,color)
-        
-        
     def show(self):
         """
         Write the current framebuf to the display
         """
-        if self._show_y_min > self._show_y_max:
-            # nothing to show
-            return
-        
         # mh_if TDECK:
-        # TDeck shares SPI with SDCard
-        self.spi.init()
+        # # TDeck shares SPI with SDCard
+        # self.spi.init()
         # mh_end_if
 
         # clamp min and max
-        if self._show_y_min < 0:
-            self._show_y_min = 0
-        if self._show_y_max > self.height:
-            self._show_y_max = self.height
+        y_min, y_max = self._reset_show_min()
+
+        if y_min >= y_max:
+            # nothing to show
+            return
 
         self._set_window(
             0,
-            self._show_y_min,
+            y_min,
             self.width - 1,
-            self._show_y_max - 1,
+            y_max - 1,
             )
         
         if self.use_tiny_buf:
-            self._write_tiny_buf()
+            self._write_tiny_buf(y_min, y_max)
         else:
-            self._write_normal_buf()
+            self._write_normal_buf(y_min, y_max)
 
-        self._reset_show_min()
+        
 
         # mh_if TDECK:
-        # TDeck shares SPI with SDCard
-        self.spi.deinit()
+        # # TDeck shares SPI with SDCard
+        # self.spi.deinit()
         # mh_end_if
-        
-    def blit_buffer(self, buffer, x, y, width, height, key=-1, palette=None):
-        """
-        Copy buffer to display framebuf at the given location.
-
-        Args:
-            buffer (bytes): Data to copy to display
-            x (int): Top left corner x coordinate
-            Y (int): Top left corner y coordinate
-            width (int): Width
-            height (int): Height
-            key (int): color to be considered transparent
-            palette (framebuf): the color pallete to use for the buffer
-        """
-        self._set_show_min(y, y + height)
-        if not isinstance(buffer, framebuf.FrameBuffer):
-            buffer = framebuf.FrameBuffer(
-                buffer, width, height,
-                framebuf.GS4_HMSB if self.use_tiny_buf else framebuf.RGB565,
-                )
-        
-        self.fbuf.blit(buffer, x, y, key, palette)
-
-
-    def rect(self, x, y, w, h, color, fill=False):
-        """
-        Draw a rectangle at the given location, size and color.
-
-        Args:
-            x (int): Top left corner x coordinate
-            y (int): Top left corner y coordinate
-            width (int): Width in pixels
-            height (int): Height in pixels
-            color (int): 565 encoded color
-        """
-        self._set_show_min(y, y + h)
-        color = self._format_color(color)
-        self.fbuf.rect(x,y,w,h,color,fill)
-
-
-    def ellipse(self, x, y, xr, yr, color, fill=False, m=0xf):
-        """
-        Draw an ellipse at the given location, radius and color.
-
-        Args:
-            x (int): Center x coordinate
-            y (int): Center y coordinate
-            xr (int): x axis radius
-            yr (int): y axis radius
-            color (int): 565 encoded color
-            fill (bool): fill in the ellipse. Default is False
-        """
-        self._set_show_min(y - yr, y + yr + 1)
-        color = self._format_color(color)
-        self.fbuf.ellipse(x,y,xr,yr,color,fill,m)
-
-
-    def fill_rect(self, x, y, width, height, color):
-        """
-        Draw a rectangle at the given location, size and filled with color.
-        
-        This is just a wrapper for the rect() method,
-        and is provided for compatibility with the original st7789py driver.
-
-        Args:
-            x (int): Top left corner x coordinate
-            y (int): Top left corner y coordinate
-            width (int): Width in pixels
-            height (int): Height in pixels
-            color (int): 565 encoded color
-        """
-        self._set_show_min(y, y + height)
-        self.rect(x, y, width, height, color, fill=True)
-
-
-    def fill(self, color):
-        """
-        Fill the entire FrameBuffer with the specified color.
-
-        Args:
-            color (int): 565 encoded color
-        """
-        # whole display must show
-        self._set_show_min(0, self.height)
-        color = self._format_color(color)
-        self.fbuf.fill(color)
-
-
-    def line(self, x0, y0, x1, y1, color):
-        """
-        Draw a single pixel wide line starting at x0, y0 and ending at x1, y1.
-
-        Args:
-            x0 (int): Start point x coordinate
-            y0 (int): Start point y coordinate
-            x1 (int): End point x coordinate
-            y1 (int): End point y coordinate
-            color (int): 565 encoded color
-        """
-        self._set_show_min(
-            min(y0,y1),
-            max(y0,y1)
-            )
-        color = self._format_color(color)
-        self.fbuf.line(x0, y0, x1, y1, color)
-
-
-    def scroll(self,xstep,ystep):
-        """
-        Shift the contents of the FrameBuffer by the given vector.
-        This may leave a footprint of the previous colors in the FrameBuffer.
-
-        Unlike vscsad which uses the hardware for scrolling,
-        this method scrolls the framebuffer itself.
-        This is a wrapper for the framebuffer.scroll method:
-        """
-        self._set_show_min(0, self.height)
-        self.fbuf.scroll(xstep,ystep)
-
-
-    @micropython.viper
-    def _bitmap_text(self, font, text, x:int, y:int, color:int):
-        """
-        Internal viper method to draw text.
-        Designed to be envoked using the 'text' method.
-
-        Args:
-            font (module): font module to use
-            text (str): text to write
-            x (int): column to start drawing at
-            y (int): row to start drawing at
-            color (int): encoded color to use for characters
-        """
-        width = int(font.WIDTH)
-        height = int(font.HEIGHT)
-        self_width = int(self.width)
-        self_height = int(self.height)
-
-        utf8_scale = height // 8
-        
-        # early return for text off screen
-        if y >= self_height or (y + height) < 0:
-            return
-        
-        glyphs = ptr8(font.FONT)
-        
-        char_px_len = width * height
-        
-        first = int(font.FIRST)
-        last = int(font.LAST)
-        
-        
-        use_tiny_fbuf = bool(self.use_tiny_buf)
-        fbuf16 = ptr16(self.fbuf)
-        fbuf8 = ptr8(self.fbuf)
-        
-        for char in text:
-            ch_idx = int(ord(char))
-            
-            # only draw chars that exist in font
-            if first <= ch_idx < last:
-                bit_start = (ch_idx - first) * char_px_len
-                
-                px_idx = 0
-                while px_idx < char_px_len:
-                    byte_idx = (px_idx + bit_start) // 8
-                    shift_amount = 7 - ((px_idx + bit_start) % 8)
-                    
-                    target_x = x + px_idx % width
-                    target_y = y + px_idx // width
-                    
-                    # dont draw pixels off the screen (ptrs don't check your work!)
-                    if ((glyphs[byte_idx] >> shift_amount) & 0x1) == 1 \
-                    and 0 <= target_x < self_width \
-                    and 0 <= target_y < self_height:
-                        target_px = (target_y * self_width) + target_x
-                        
-                        # I tried putting this if/else before px loop,
-                        # surprisingly, there was not a noticable speed difference,
-                        # and the code was harder to read. So, I put it back.
-                        if use_tiny_fbuf:
-                            # pack 4 bits into 8 bit ptr
-                            target_idx = target_px // 2
-                            dest_shift = ((target_px + 1) % 2) * 4
-                            dest_mask = 0xf0 >> dest_shift
-                            fbuf8[target_idx] = (fbuf8[target_idx] & dest_mask) | (color << dest_shift)
-                        else:
-                            # draw to 16 bits
-                            target_idx = target_px
-                            fbuf16[target_idx] = color
-                    
-                    px_idx += 1
-                x += width
-            else:
-                # try drawing with utf8 instead
-                x += int(self.utf8_putc(ch_idx, x, y, color, utf8_scale))
-
-            # early return for text off screen
-            if x >= self_width:
-                return
-
-
-    @micropython.viper
-    def utf8_putc(self, char:int, x:int, y:int, color:int, scale:int) -> int:
-        """Render a single character on the screen."""
-        width = 4 if char < 128 else 8
-        height = 8
-
-        if not 0x0000 <= char <= 0xFFFF:
-            return width * scale
-        
-        # set up viper variables
-        use_tiny_fbuf = bool(self.use_tiny_buf)
-        fbuf16 = ptr16(self.fbuf)
-        fbuf8 = ptr8(self.fbuf)
-        self_width = int(self.width)
-        self_height = int(self.height)
-
-        # calculate the offset in the binary data
-        offset = char * 8
-
-        # mh_if frozen:
-        # # Read the font data directly from the memoryview
-        # cur = ptr8(utf8)
-        # mh_else:
-        # seek to offset and read 8 bytes
-        self.utf8_font.seek(offset)
-        cur = ptr8(self.utf8_font.read(8))
-        # mh_end_if
-
-        # y axis is inverted - we start from bottom not top
-        y += (height - 1) * scale - 1
-
-        # iterate over every character pixel
-        px_idx = 0
-        max_px_idx = width * height
-        while px_idx < max_px_idx:
-            # which byte to fetch from the ptr8,
-            # and how far to shift (to get 1 bit)
-            ptr_idx = px_idx // 8
-            shft_idx = px_idx % 8
-            # mh_if frozen:
-            # # if reading from memoryview, add offset now
-            # ptr_idx += offset
-            # mh_end_if
-
-            # calculate x/y position from pixel index
-            target_x = x + ((px_idx % width) * scale)
-            target_y = y - ((px_idx // width) * scale)
-            
-            if (cur[ptr_idx] >> shft_idx) & 1 == 1:
-                # iterate over x/y scale
-                scale_idx = 0
-                num_scale_pixels = scale * scale
-                while scale_idx < num_scale_pixels:
-                    xsize = scale_idx % scale
-                    ysize = scale_idx // scale
-
-                    target_px = ((target_y + ysize) * self_width) + target_x + xsize
-                    if 0 <= (target_x + xsize) < self_width \
-                    and 0 <= (target_y + ysize) < self_height:
-                        if use_tiny_fbuf:
-                            # pack 4 bits into 8 bit ptr
-                            target_idx = target_px // 2
-                            dest_shift = ((target_px + 1) % 2) * 4
-                            dest_mask = 0xf0 >> dest_shift
-                            fbuf8[target_idx] = (fbuf8[target_idx] & dest_mask) | (color << dest_shift)
-                        else:
-                            # draw to 16 bits
-                            target_idx = target_px
-                            fbuf16[target_idx] = color
-                    scale_idx += 1
-            px_idx += 1
-        
-        # return x offset for drawing next char
-        return width * scale
-
-
-    @micropython.viper
-    def _utf8_text(self, text, x:int, y:int, color:int):
-        """Draw text, including utf8 characters"""
-        str_len = int(len(text))
-        
-        idx = 0
-        while idx < str_len:
-            char = text[idx]
-            ch_ord = int(ord(char))
-            if ch_ord >= 128:
-                x += int(self.utf8_putc(ch_ord, x, y, color, 1))
-            else:
-                self.fbuf.text(char, x, y, color)
-                x += 8
-            idx += 1
-        
-
-    def text(self, text, x, y, color, font=None):
-        """
-        Draw text to the framebuffer.
-        
-        Text is drawn with no background.
-        If 'font' is None, uses the builtin framebuffer font.
-        
-        Args:
-            text (str): text to write
-            x (int): column to start drawing at
-            y (int): row to start drawing at
-            color (int): encoded color to use for text
-            font (optional): bitmap font module to use
-        """
-        color = self._format_color(color)
-
-
-        if font:
-            self._set_show_min(y, y + font.HEIGHT)
-            self._bitmap_text(font, text, x, y, color)
-        else:
-            self._set_show_min(y, y + 8)
-            self._utf8_text(text, x, y, color)
-
-
-    def bitmap(self, bitmap, x, y, index=0, key=-1, palette=None):
-        """
-        Draw a bitmap on display at the specified column and row
-
-        Args:
-            bitmap (bitmap_module): The module containing the bitmap to draw
-            x (int): column to start drawing at
-            y (int): row to start drawing at
-            index (int): Optional index of bitmap to draw from multiple bitmap
-                module
-            key (int): colors that match the key will be transparent.
-        """
-        if self.width <= x or self.height <= y:
-            return
-        
-        if palette is None:
-            palette = bitmap.PALETTE
-        
-        self._bitmap(bitmap, x, y, index, key, palette)
-
-
-    @micropython.viper
-    def _bitmap(self, bitmap, x:int, y:int, index:int, key:int, palette):
-        
-        width = int(bitmap.WIDTH)
-        height = int(bitmap.HEIGHT)
-        self_width = int(self.width)
-        self_height = int(self.height)
-        
-        palette_len = int(len(palette))
-        bpp = int(bitmap.BPP)
-        bitmap_pixels = height * width
-        starting_bit = bpp * bitmap_pixels * index  # if index > 0 else 0
-        
-        use_tiny_buf = bool(self.use_tiny_buf)
-        
-        self._set_show_min(y, y + height)
-        
-        
-        # format color palette into a pointer
-        palette_buf = bytearray(palette_len * 2)
-        palette_ptr = ptr16(palette_buf)
-        for i in range(palette_len):
-            palette_ptr[i] = int(
-                self._format_color(palette[i])
-                )
-        key = int(self._format_color(key))
-        
-        bitmap_ptr = ptr8(bitmap._bitmap)
-        fbuf8 = ptr8(self.fbuf)
-        fbuf16 = ptr16(self.fbuf)
-        
-        bitmask = 0xffff >> (16 - bpp)
-        
-        # iterate over pixels
-        px_idx = 0
-        while px_idx < bitmap_pixels:
-            source_bit = (px_idx * bpp) + starting_bit
-            source_idx = source_bit // 8 
-            source_shift = 7 - (source_bit % 8)
-            
-            # bitmap value is an index in the color palette
-            source = (bitmap_ptr[source_idx] >> source_shift) & bitmask
-            clr = palette_ptr[source]
-            
-            target_x = x + px_idx % width
-            target_y = y + px_idx // width
-            
-            # dont draw pixels off the screen (ptrs don't check your work!)
-            if clr != key \
-            and 0 <= target_x < self_width \
-            and 0 <= target_y < self_height:
-                
-                # convert px coordinate to an index
-                target_px = (target_y * self_width) + target_x
-                
-                if use_tiny_buf:
-                    # writing 4-bit pixels
-                    target_idx = target_px // 2
-                    dest_shift = ((target_px + 1) % 2) * 4
-                    dest_mask = 0xf0 >> dest_shift
-                    fbuf8[target_idx] = (fbuf8[target_idx] & dest_mask) | (clr << dest_shift)
-                else:
-                    # TODO: TEST THIS! (has only been tested for tiny fbuf)
-                    # writing 16-bit pixels
-                    target_idx = target_px
-                    fbuf16[target_idx] = clr
-            
-            px_idx += 1
-
-
-    def polygon(self, coords, x, y, color, fill=False):
-        """
-        Draw a polygon from an array of coordinates
-
-        Args:
-            coords (array('h')): An array of x/y coordinates defining the shape
-            x (int): column to start drawing at
-            y (int): row to start drawing at
-            color (int): Color of polygon
-            fill (bool=False) : fill the polygon (or draw an outline)
-        """
-        # calculate approx height so min/max can be set
-        h = max(coords)
-        self._set_show_min(y, y + h)
-        color = self._format_color(color)
-        self.fbuf.poly(x, y, coords, color, fill)
